+++ core/state/app/model.go
@@ -4,7 +4,7 @@ import "math/big"
 
 type Model struct {
 	TotalSlashed *big.Int
-	CoinsCount   uint64
+	CoinsCount   uint32
 	MaxGas       uint64
 
 	markDirty func()
@@ -36,11 +36,11 @@ func (model *Model) setTotalSlashed(totalSlashed *big.Int) {
 	model.TotalSlashed = totalSlashed
 }
 
-func (model *Model) getCoinsCount() uint64 {
+func (model *Model) getCoinsCount() uint32 {
 	return model.CoinsCount
 }
 
-func (model *Model) setCoinsCount(count uint64) {
+func (model *Model) setCoinsCount(count uint32) {
 	if model.CoinsCount != count {
 		model.markDirty()
 	}
